+++ bindings/rust/libgpiod/src/lib.rs
@@ -193,7 +193,7 @@ pub mod line {
                         errno::errno(),
                     ))
                 }
-                _ => return Err(Error::InvalidEnumValue("Value", val as i32)),
+                _ => return Err(Error::InvalidEnumValue("Value", val)),
             })
         }
